+++ docs/conf.py
@@ -26,7 +26,6 @@ extensions = [
     'guzzle_sphinx_theme',
     'nbsphinx',
     'numpydoc',
-    'nengo.utils.docutils',
 ]
 
 # -- sphinx.ext.autodoc