+++ CHANGES.rst
@@ -19,8 +19,8 @@ Release History
    - Removed
    - Fixed
 
-2.7.1 (unreleased)
-==================
+2.8.0 (June 9, 2018)
+====================
 
 **Added**